pyproject.toml
[project]
name = "gbif"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ichatbio-sdk>=0.2.2",
    "instructor>=1.9.2",
    "openai>=1.97.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "tenacity>=9.1.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --import-mode=importlib"
